Deep overhaul of SSS implementation (#1272)
* Deep overhaul of SSS implementation

* Add comment

* Fix directional dipole

* Fix Linux build

* Simplify test scene

* Update directional dipole reference images

* Fix warnings on Windows

* Formatting and cosmetic tweaks

* Fix Linux build

* Skip Fresnel computations if Fresnel weight is 0

+++ src/appleseed/foundation/meta/benchmarks/benchmark_microfacet.cpp
@@ -70,7 +70,7 @@ BENCHMARK_SUITE(Foundation_Math_Microfacet)
 
         void evaluate(const float alpha_x, const float alpha_y)
         {
-            Vector2d s;
+            Vector2f s;
             s[0] = rand_float2(m_rng);
             s[1] = rand_float2(m_rng);
